Switch to :require-ssl and command checks
Rolled back all of the attempts to configure the
server since they didn't seem compatible with our
other scripts. There may be FTP clients they could
have worked with, but research shows that FTP server
protocols are a bit unclear as to how this should
work.

Instead, i've designated a few commands as ones 
that the server will check for a secure channel
before accepting if you set `:require-ssl` to
true. Namely, USER and PASS. If either of those
are sent before AUTH TLS is established, the server
will disconnect.

This may have the side effect of making more attacks
fail faster, as attackers likely won't know and won't
issue AUTH TLS before attempting to log in with
fake accounts.